requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-autocomplete-light==3.8.2
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-import-export==2.5.0
django-mathfilters==1.0.0
django-pwa==1.0.10
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
reportlab==3.5.66
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
webencodings==0.5.1
whitenoise==4.1.3
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
pip~=21.0.1
setuptools~=41.2.0
djangorestframework~=3.12.4